// frontend/src/api/displaySettings.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "./graphql";

export type DisplaySettingsVariables = {
  input: {
    logo: string;
    theme: string;
  };
};

export type DisplaySettingsResult = {
  displaySettings: {
  customTheme: {
  value: string;
  hash: string;
} | null;
  customLogo: {
  value: string;
  hash: string;
} | null;
};
};

export const DisplaySettings = {
  query: "query displaySettings($input: DisplaySettingsHash!) {\n  displaySettings(input: $input) {\n    customTheme {\n      value\n      hash\n    }\n    customLogo {\n      value\n      hash\n    }\n  }\n}",
} as TypedDocument<DisplaySettingsResult, DisplaySettingsVariables>;

export type UpdateDisplaySettingsVariables = {
  input: {
    customLogo?: string | null;
    customTheme?: string | null;
  };
};

export type UpdateDisplaySettingsResult = {
  updateDisplaySettings: ({
  __typename: "UpdateResult";
} & {
  theme: string | null;
  logo: string | null;
}) | ({
  __typename: "UpdateDisplaySettingsError";
} & {
  error: string;
});
};

export const UpdateDisplaySettings = {
  query: "mutation updateDisplaySettings($input: DisplaySettingsInput!) {\n  updateDisplaySettings(input: $input) {\n    ... on UpdateResult {\n      __typename\n      theme\n      logo\n    }\n    ... on UpdateDisplaySettingsError {\n      __typename\n      error\n    }\n  }\n}",
} as TypedDocument<UpdateDisplaySettingsResult, UpdateDisplaySettingsVariables>;